refactor(telemetry): rename prober.rs to el_prober.rs
